Update index.html
Updated the dropdown to Drowsy Chaperone block so that it has an ellipse button. The button opens the pre-existing modal that now is filled with data about the big three and cast. *That will need to be changed once XML is learned as it takes up at least 20% of the lines in the index.html file. As well changed the information provided in the well underneath the main block so that it contains times and dates of the show, as well as where to reserve tickets.